Cleaned up the code and made it black compatible and pylint 10/10 rated.
